Google People (Contact) - Create Contact (#5000)
* create-contact action

* updates

* additionalProps optional for create-contact
