+++ components/google_contacts/actions/delete-contact/delete-contact.mjs
@@ -5,7 +5,7 @@ export default {
   key: "google_contacts-delete-contact",
   name: "Delete Contact",
   description: "Deletes a contact. [See the docs here](https://developers.google.com/people/api/rest/v1/people/deleteContact)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     ...common.props,
@@ -23,7 +23,7 @@ export default {
       });
     },
     emitSummary($) {
-      $.export("$summary", "Successfully deleted contact");
+      $.export("$summary", `Successfully deleted contact with ID ${this.resourceName.split("/").pop()}`);
     },
   },
 };
