+++ components/google_contacts/actions/list-contacts/list-contacts.mjs
@@ -6,7 +6,7 @@ export default {
   key: "google_contacts-list-contacts",
   name: "List Contacts",
   description: "Lists all contacts of the authenticated user. [See the docs here](https://developers.google.com/people/api/rest/v1/people.connections/list)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     ...common.props,
