+++ components/google_contacts/actions/update-contact/update-contact.mjs
@@ -5,7 +5,7 @@ export default {
   key: "google_contacts-update-contact",
   name: "Update Contact",
   description: "Updates a contact. [See the docs here](https://developers.google.com/people/api/rest/v1/people/updateContact)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   props: {
     ...common.props,
@@ -24,126 +24,25 @@ export default {
     },
   },
   async additionalProps() {
-    const props = {};
-    if (this.updatePersonFields.includes("names")) {
-      props.firstName = {
-        type: "string",
-        label: "First Name",
-        description: "Contact's first name",
-      };
-      props.middleName = {
-        type: "string",
-        label: "Middle Name",
-        description: "Contact's middle name",
-      };
-      props.lastName = {
-        type: "string",
-        label: "Last Name",
-        description: "Contact's last name",
-      };
-    }
-    if (this.updatePersonFields.includes("emailAddresses")) {
-      props.email = {
-        type: "string",
-        label: "Email Address",
-        description: "Contact's email address",
-      };
-    }
-    if (this.updatePersonFields.includes("phoneNumbers")) {
-      props.phoneNumber = {
-        type: "string",
-        label: "Phone Number",
-        description: "Contact's phone number",
-      };
-    }
-    if (this.updatePersonFields.includes("addresses")) {
-      props.streetAddress = {
-        type: "string",
-        label: "Street Address",
-        description: "Contact's street address",
-      };
-      props.city = {
-        type: "string",
-        label: "City",
-        description: "Contact's city",
-      };
-      props.state = {
-        type: "string",
-        label: "State",
-        description: "Contact's state/region",
-      };
-      props.zipCode = {
-        type: "string",
-        label: "Zip Code",
-        description: "Contact's zip code",
-      };
-      props.country = {
-        type: "string",
-        label: "Country",
-        description: "Contact's country",
-      };
-    }
-    return props;
+    return this.getPersonFieldProps(this.updatePersonFields);
   },
   methods: {
+    ...common.methods,
     async processResults(client) {
       const {
-        resourceName,
         updatePersonFields,
-        firstName,
-        middleName,
-        lastName,
-        email,
-        phoneNumber,
-        streetAddress,
-        city,
-        state: region,
-        zipCode: postalCode,
-        country,
+        resourceName,
       } = this;
+
+      const requestBody = this.getRequestBody(updatePersonFields, this);
 
       const { etag } = await this.googleContacts.getContact(client, {
         resourceName,
         personFields: "names",
       });
 
-      const requestBody = {
-        etag,
-      };
-      if (updatePersonFields.includes("addresses")) {
-        requestBody.addresses = [
-          {
-            streetAddress,
-            city,
-            region,
-            postalCode,
-            country,
-          },
-        ];
-      }
-      if (updatePersonFields.includes("emailAddresses")) {
-        requestBody.emailAddresses = [
-          {
-            value: email,
-          },
-        ];
-      }
-      if (updatePersonFields.includes("names")) {
-        requestBody.names = [
-          {
-            familyName: lastName,
-            middleName,
-            givenName: firstName,
-          },
-        ];
-      }
-      if (updatePersonFields.includes("phoneNumbers")) {
-        requestBody.phoneNumbers = [
-          {
-            value: phoneNumber,
-          },
-        ];
-      }
+      requestBody.etag = etag;
+
       return this.googleContacts.updateContact(client, {
         resourceName,
         requestBody,
@@ -151,7 +50,7 @@ export default {
       });
     },
     emitSummary($) {
-      $.export("$summary", "Successfully updated contact");
+      $.export("$summary", `Successfully updated contact with ID ${this.resourceName.split("/").pop()}`);
     },
   },
 };
