+++ components/google_contacts/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pipedream/google_contacts",
-  "version": "0.0.2",
+  "version": "0.0.1",
   "description": "Pipedream Gooogle Contacts Components",
   "main": "google_contacts.app.mjs",
   "keywords": [
